+++ src/org/ojalgo/matrix/decomposition/HermitianEvD.java
@@ -410,8 +410,7 @@ public abstract class HermitianEvD<N extends Number> extends EigenvalueDecomposi
 
     @Override
     protected MatrixStore<N> makeD() {
-        final DiagonalBasicArray<Double> diagonal = new DiagonalBasicArray<>(Primitive64Array.wrap(d), null, null, ZERO);
-        return this.wrap(diagonal).diagonal().get();
+        return this.makeDiagonal(Primitive64Array.wrap(d)).get();
     }
 
     @Override